Add readDir tests
